234.2 Initial App Setup(Section 18: Async Actions with Redux Thunk)
So inside the sarf see folder I'll begin by creating an index file.

I'm going to import re-act from re-act import reac Dom from re-act dash Dom.

I'm going to get the provider element from reactor's docs and then I'll make sure that I also get create

store from the redux library itself.

And then after that I'll also assume that I'm going to eventually have an app component we haven't created

just yet but we will in just a moment.

So import app from components app and then I'm also going to assume that we're going to eventually have

a couple of reducers inside this application so I'm going to import reducers from reducers and then

at the bottom we'll do our re-act dum dum render a call so I'm going to show my provider tag I'm going

to pass it reate store and when I call create store right here I'll be sure to pass in all my different

reducers I'll close off the provider tag and then inside there I'll show my app component.

And then finally as the second argument to the re-act Dannatt render a call we'll do the same thing

we've always done before.

We'll tell re-act where it needs to render our application.

So I'll say document Dom queries selector Halland route like the.

All right.

So just some typical boilerplate that you're going to see in just about every single redux application.

So now at this point if we flip back over to our browser we'll very quickly see an error message because

we do not have a app file.

We also do not have any reducers just yet.

So it's very quickly make a kind of default little app component and some default little reducers just

to get our application working and appearing on the screen.

So inside my SIRC directory I'll make a new folder called component's inside of component's I'll make

my apt.

Genius file.

And then inside of here I will import re-act from re-act I'll create my app component as a functional

component and right now we'll just have it return a div that has that text app.

Now we've seen many times so far throughout this course as we've been making use of semantic UI.

We've always had to come back to this app component and add in that class name of UI container.

Remember that class name when used with semantic UI gets us the nice margin on the left hand side and

the right hand side.

So we're going to make sure that we just go ahead and apply that class name right now.

So we don't have to come back and do it later as we've done throughout this course so far.

So on that div I'll put on a class name of UI container and then finally at the bottom I will export

default app like so.

All right so now if we save this we can flip back over and we're going to see that the error message

has now been updated.

Now it says that it cannot resolve reducers which means essentially we don't have a reducers directory.

We don't have any reducers to make use of whatsoever.

Now we've already done a lot of typing inside this video.

So let's take a quick pause right here.

When we come back the next section will start creating some reducers for this application just to get

our application rendering and working inside the browser.