235.1. Tricking Redux with Dummy Reducers(Section 18: Async Actions with Redux Thunk)
In less video we added in a ton of boilerplate to our project.

But as we saw we're still getting this air around the fact that we do not have reducers folder or an

index such as file inside there.

So let's take care of that very quickly.

Back inside my SIRC folder I'll make a new folder called reducers and inside of there I'll make a new

file called the index such as file.

So as we've seen previously we can import combine reducers from redux and then we can export default

combine reducers and we'll pass that thing and object.

Now at present we have nothing to put inside this object because we don't have any reducers inside of

application.

Sure we'll talk about what reducers we need in just a moment but we don't have any reducer whatsoever

to put in here.

As it stands.

So let's just try saving the file as is and see if our application works now.

I flip back over I'll see that while my text is displaying here on the screen.

You know I see that Aztec's to appear.

I can open up my console and unfortunately I'm getting a little kind of warning message here.

Really I would call this an error not a warning.

It only appears to be kind of a warning because our application still appeared on the screen.

But this definitely is an error.

So the air here says that our store does not have a valid reducer.

And hey that totally makes sense.

We called Combined reducers but we don't have any reducers to pass to it just yet.

So it turns out that this happens rather frequently when you're first getting started on any given redux

project.

You might start up the project put together some boilerplate but you don't know what reducers you're

going to need.

So you just kind of want get your application working and get that error message to go away.

You might eventually come back and add in some reducers which would solve the problem but maybe for

right now we don't really want to work on the reducers inside of our app at all.

Maybe we want to work on the Riak side of the project or something like that.