235.2. Tricking Redux with Dummy Reducers(Section 18: Async Actions with Redux Thunk)
So as a little work around there's a tiny little trick that we can do here.

I would barely even call it a trick.

Inside the combined reducers call I can put in some dummy key like some dummy variable you can be total

jibberish like that if I want to it can be literally the word dummy or even something like replace me.

And then we can assign a dummy key an arrow function that is always going to return some fixed value.

Like maybe the number 10 or the number 9 9 9 or the string.

Hi there.

There's no purpose to this thing right here.

It's just tricking redux into thinking that it has a valid reducer.

That's it.

And the only reason we're doing this is to get that error message to go away.

Like I said anytime you start up a redux project you're probably going to want to do something like

this very briefly until you get a really good idea of how to structure your reducers in what reducers

you're going to need.

So now that I've got the dummy producer inside of here I can flip back over the error message has gone

away.

So now we are free to work on all the different other aspects of our application such as enhancing some

Riak components or maybe creating some action creators or whatever it might be.

And we don't have to worry about the fact that redux is just refusing to play ball.

So now we've kind of got some leverage here and we could start working on anything that we want now

to be honest Republican to start working on the reducers first but hey that's just life.

I just want to show you this little thing that you can do to get that your message to go away.

All right let's take another quick pause right here and we'll continue in the next section.