236. A Touch More Setup(Section 18: Async Actions with Redux Thunk)
In this video we're going to continue doing a little bit of low level set up on our project in particular.

We're going to start putting together our post list component once we've got this component put together

and shown inside the app will then start to call an action creator to go off and fetch the data we care

about and will start talking about some interesting data fetch stuff around redux.

All right so let's get to it.

I can cite my editor.

I'm going to find the components directory and inside there I'm going to create a new file called the

post list G-S file.

Then inside of your I'm going to import re-act from react and I would make this a class based component.

You'll see why in just a moment.

So the find my render method I will return a div with a text post list and then at the bottom I'll do

an export default lowest list like so next step I'm going to take my post lists and import it into the

app and show it inside the component.

So inside the app I will import post list from Host List and then finally I'll replace the text right

there with the post list I'm going to make this a multi-line GSX expression.

So parentheses and closing parentheses.

And don't forget to clean up that trailing semicolon at the end.

All right so let's test this out really quickly.

I flip back over to the application.

I see pulseless on the screen.

All right.

Lesing I want to do as a little bit of setup is to wire up the semantic UI library or the CSSA library.

One more time to remember to open up a new tab will search Google for semantic UI Cdn or find the Cdn

link or search for semantic men CSSA on this page.

There it is right there saw a copy of that link and then as usual we'll open up our public index.

H.M. file right underneath the MedTech I'll put a link in here with a rel style sheet and an H.F. of

that link will close a tag off on the far right hand side.

All right.

So if I flip back over everything should look pretty nice.

I see the new font appear on the screen.

OK.

So let's pause right here.

When we come back the next section we're going to start talking about the different methodologies that

we're going to use to eventually fetch some data inside of redux.

So we'll see you in just a minute.