244.2. Shortened Syntax with Redux Thunk(Section 18: Async Actions with Redux Thunk)
But whenever you make use of redux thunk you will very commonly seen this.

See this thing refactor just a little bit.

So first off it's very common to return a arrow function instead of a keyword function like so.

Secondly you'll notice that you're not using this state argument or function anywhere inside the function.

So we don't have to always list out get state we are making use of dispatch.

So we will leave that in.

Remember that whenever we have a aero function with only one arguments the parentheses around it are

optional.

So I can remove those and then finally on the outer function you'll notice that we do not have any other

variables inside of here.

We just have one expression that we are immediately returning and so on the outer function we can optionally

remove those semi-colons or mean Curly braces at the top and the bottom and remove the return key word

right there.

And so this entire statement can be shortened up to just this like so totally equivalent to what we

just had a moment ago.

We just took made use of a couple of yes 20:15 pieces of syntax that's all.

So the syntax you see right here.

You are going to very frequently see in a lot of redux projects.

Remember all this is doing is we're defining a function that is going to return a function that says

pretty much it.

This is what we're going to end up with right here.

So I'm going to save this and then I'm just going to flip back over to my browser really quick and make

sure that we do not have any error messages back over here I don't see any error messages so it looks

like we're good to go.

All right so now that we understand how to create an asynchronous action creators we're not going to

go back to our original flow around loading data up inside of a redux application.

So this one right here.

So some steps in this flow are going to kind of change now that we understand how to create asynchronous

action creators but essentially at this point we have now taken care of those steps right there and

now it's really just up to us to define a reducer that's going to see the action we are dispatching

and get the data from that API off the payload property.

So let's take a quick pause right here and we'll start taking care of our inducers in the next video.