246.3. Return Values from Reducers(Section 19: Redux Store Design)
It's when I failed to place a return statement or I specifically return a value of undefined that we're

going to run into that error message.

Now if you look at this message very closely it says returned undefined during initialization.

So when redux first boots up it's going to run you of your reducers one time and we'll dive into that

process a bit more closely in just a moment.

But essentially when redux first boots up it's going to call your reducers one time.

And so right you're this during initialization we're right there.

That means that the first time you're douceur was ran it returned the value undefined.

Now the rule around not returning undefined is not only about initialization your reducer can never

ever return undefined whether it is an initialization or at some point in time the future when an action

has been dispatched.

OK so that's rule number one around reducers so we must always return any value besides undefined.

So let's take a quick pause right here.