253.2. Switch Statements in Reducers(Section 19: Redux Store Design)
We would instead have a switch statement a switch statement is valid javascript syntax this is not react

specific it's not redux specific.

For the switch we're going to take a look at the action duct tape type property and then we will define

a couple of cases inside of here.

So I'm going to first define a case of fetch oats and then I'll put a colon after that and then underneath

that case I'm going to put down what I want to do.

If the action not type property is equal to the string fetch underscore posts.

If that is the case then I want to return action payload.

So in practice if we end up having a lot of types that a single reducer cares about you'll end up seeing

a lot of case statements so you'll see like one case you'll see another case for another type and another

case for another type.

And we just add them all into this switch statement.

In this case producer is only ever going to care about the type fecche posts.

So we don't have to worry about adding in any other case statements.

Now let's make sure that we always return some value from our reducer.

We're always going to add in a default case to our switch statement.

So if we do not match the case of fecche posts once said fall through to the default Keystone here and

in the default case the case in which we do not have a action type that matches fetch posts are going

to simply return state like so.

All right so what you see right here is extremely common syntax to see inside of reducers the if statements.

We were working with before.

Definitely not very common.

I just kept it very simple so you could very easily see what was going on.

It is much more common to see inside of reducer a switch statement that looks exactly like this.

Although frankly you will have many other cases besides just the one that we have.

OK.

So this is looking pretty good.

So let's take a pause right here and we'll continue in the next video.